Added some basic visual components, added icon lib, and separated miscellaneous CSS.
