+++ package.json
@@ -12,7 +12,8 @@
     "next": "10.0.6",
     "react": "17.0.1",
     "react-bootstrap": "^1.4.3",
-    "react-dom": "17.0.1"
+    "react-dom": "17.0.1",
+    "react-icons": "^4.1.0"
   },
   "devDependencies": {
     "@types/node": "^14.14.22",
